# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Water Meter Backend API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /meters:
    get:
      operationId: listMeters
      tags: [meters]
      summary: List all meters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        "200":
          description: List of meters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMeter
      tags: [meters]
      summary: Register a new meter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeterRequest"
      responses:
        "201":
          description: Meter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Meter serial already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meters/{id}:
    get:
      operationId: getMeter
      tags: [meters]
      summary: Get meter by ID
      parameters:
        - $ref: "#/components/parameters/MeterId"
      responses:
        "200":
          description: Meter details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
        "404":
          description: Meter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meters/{id}/status:
    get:
      operationId: getMeterStatus
      tags: [meters]
      summary: Get meter real-time status
      parameters:
        - $ref: "#/components/parameters/MeterId"
      responses:
        "200":
          description: Meter status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterStatus"
        "404":
          description: Meter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meters/{id}/consumption:
    get:
      operationId: getMeterConsumption
      tags: [meters]
      summary: Get meter consumption history
      parameters:
        - $ref: "#/components/parameters/MeterId"
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: daily
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Consumption data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionResponse"
        "404":
          description: Meter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    MeterId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Meter:
      type: object
      required: [id, serialNumber, status, ownerId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        serialNumber:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        type:
          type: string
          enum: [residential, commercial, industrial]
        installDate:
          type: string
          format: date
          nullable: true
        location:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        ownerId:
          type: string
          format: uuid
        deviceId:
          type: string
          nullable: true
        lastReading:
          type: number
          format: double
          nullable: true
        lastReadingAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MeterStatus:
      type: object
      required: [meterId, online, valveState]
      properties:
        meterId:
          type: string
          format: uuid
        online:
          type: boolean
        lastSeen:
          type: string
          format: date-time
          nullable: true
        batteryLevel:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        signalStrength:
          type: integer
          nullable: true
          description: RSSI signal strength
        valveState:
          type: string
          enum: [open, closed, unknown]
        currentReading:
          type: number
          format: double
          nullable: true
        firmwareVersion:
          type: string
          nullable: true
    ConsumptionResponse:
      type: object
      required: [meterId, period, data]
      properties:
        meterId:
          type: string
          format: uuid
        period:
          type: string
          enum: [daily, weekly, monthly]
        totalConsumption:
          type: number
          format: double
        data:
          type: array
          items:
            type: object
            required: [date, value]
            properties:
              date:
                type: string
                format: date
              value:
                type: number
                format: double
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        details:
          type: object
          additionalProperties: true
          nullable: true
...
    CreateMeterRequest:
      type: object
      required: [serialNumber, ownerId]
      properties:
        serialNumber:
          type: string
        type:
          type: string
          enum: [residential, commercial, industrial]
          default: residential
        installDate:
          type: string
          format: date
          nullable: true
        location:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        ownerId:
          type: string
          format: uuid
        deviceId:
          type: string
          nullable: true
          description: ESP32 device identifier
...
    MeterListResponse:
      type: object
      required: [meters, total, page, limit]
      properties:
        meters:
          type: array
          items:
            $ref: "#/components/schemas/Meter"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...